test: fix data race
